+++ pkg/agent_local/ssh.go
@@ -4,12 +4,13 @@ import (
 	"context"
 
 	"goauthentik.io/platform/pkg/agent_local/types"
+	"goauthentik.io/platform/pkg/shared/grpc"
 	sshagent "goauthentik.io/platform/pkg/ssh_agent"
 )
 
 func (a *Agent) startSSH() {
 	l := a.log.WithField("logger", "agent.ssh")
-	vgrpc := &sshagent.MethodCaller{}
+	vgrpc := &grpc.MethodCaller{}
 	a.setupGRPCServer(vgrpc)
 	ag, err := sshagent.New(l, a.tr, context.Background(), vgrpc)
 	if err != nil {
