+++ pkg/shared/grpc/method_caller.go
@@ -1,4 +1,4 @@
-package sshagent
+package grpc
 
 import (
 	"context"
